Basic Networked CharacterSelectScreen

+++ Assets/Code/src/Runtime/Networking/Backends/NetworkClientPlayer.cs
@@ -5,11 +5,11 @@ namespace HouraiTeahouse.FantasyCrescendo.Networking {
 public class NetworkClientPlayer {
 
   public readonly INetworkConnection Connection;
-  public readonly uint PlayerID;
+  public readonly byte PlayerID;
   public bool IsReady;
   public PlayerConfig Config;
 
-  public NetworkClientPlayer(INetworkConnection connection, uint playerId) {
+  public NetworkClientPlayer(INetworkConnection connection, byte playerId) {
     Connection = connection;
     PlayerID = playerId;
     IsReady = false;
